move everything intro ./src dir
