+++ src/fb-assets.js
